Added lite server
