+++ index.html
@@ -7,8 +7,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
     <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-    <link rel="stylesheet" href="./assets/style/style.css">
-    <script src="assets/js/main.js" defer></script>
+    <link rel="stylesheet" href="style.css">
+    <script src="modules/main.js" defer type="module"></script>
 </head>
 
 <body>
