+++ modules/main.js
@@ -3,8 +3,10 @@ const authorInput = document.querySelector('.author');
 const form = document.querySelector('form');
 const booksListElement = document.querySelector('.listOfBooks');
 const navigationLinks = document.querySelectorAll('.nav-link');
+import { BookElement } from "./BookELement";
 
-class Main {
+
+class Main extends BookElement {
   constructor() {
     this.listOfBooks = [];
   }
@@ -14,23 +16,6 @@ class Main {
     localStorage.setItem('books', listInJson);
   };
 
-  bookElement = (title, author, id) => {
-    const element = document.createElement('li');
-    element.id = id;
-    element.innerHTML = `
-  
-    <div class="">
-        <p>${title} by ${author}</p>
-    </div>
-    <button>
-        Delete
-        <i class='bx bx-trash'></i>
-    </button>
-
-        `;
-    element.querySelector('button').addEventListener('click', () => this.removeElement(id));
-    return element;
-  };
 
   removeElement = (id) => {
     this.listOfBooks = this.listOfBooks.filter((el) => el.id !== id);
@@ -79,6 +64,7 @@ class Main {
     });
   }
 
+
   switchPages = (page) => {
     const findActivePage = document.querySelector('.opened');
     findActivePage.classList.remove('opened');
@@ -87,11 +73,13 @@ class Main {
     newActivePage.classList.remove('hidden');
     newActivePage.classList.add('opened');
   }
+
 }
 
 const mainClass = new Main();
 mainClass.formSubmit();
 mainClass.loadData();
+
 navigationLinks.forEach((el) => {
   el.addEventListener('click', () => {
     document.querySelector('.active').classList.remove('active');
